requirements.txt
flask
google-generativeai
gunicorn
pandas
requests
